Changes to Unit Class
Created a static class BaseStaticValues that has variables for things
that don't change in the game or only needs one instance of it in the
game.

Removed the variables that  were not needed in the Unit classes because
they were placed in the BaseStaticValues.

Since the Data for the each player is static, the values are tied to the
playerMech and PlayerPlane  Class

Prepared for adding in the other functions for the units that have not
been place in the Unit Class