Targeting, Shooting, Aiming, InfantryV2 and JeepV2
I got the Targeting, Shooting, and Aiming working.  I fixed the
animation for the Infantry Unit, and the jeep prefab.  All of the units
I worked on are in my Units folder.  I took one of the shot prefabs and
made a new one that has the UnitController being used.  I just called it
Shot. The other units besides the InfantryV2 and the JeepV2  will need a
few changes to get them to work with everything.